Ordenar y definir widget reutilizables
