+++ 07_Proyectos/flutter_chat_firebase/lib/src/widgets/app_icon.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class AppIcon extends StatelessWidget {
+ @override
+ Widget build(BuildContext context) {
+   //Propiedad para el titulo mas el icono a lado.
+  return Row(
+      children: <Widget>[
+        Image.asset('images/logo.png'),
+        Text('Flutter Chat',
+            style: TextStyle(
+              fontSize: 45.0,
+              fontWeight: FontWeight.w700
+            )
+        )
+      ]
+    );
+ }
+}
